+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.2.8.2
 
 -   Fix when restarting, add command to not using system jna
+-   Fix issue with installing Fabric not downloading all libraries needed